Handle nil wiki content

+++ lib/tasks/send_digests.rake
@@ -5,10 +5,12 @@ task :send_digests => :environment do
       :order => 'wiki_contents.version' )
     data = {}
     digests.each do |digest|
-      data[digest.wiki_content.project] ||= {}
-      data[digest.wiki_content.project][digest.wiki_content.page] ||= []
-      data[digest.wiki_content.project][digest.wiki_content.page].push digest.wiki_content
-      digest.destroy
+      unless digest.wiki_content.nil?
+        data[digest.wiki_content.project] ||= {}
+        data[digest.wiki_content.project][digest.wiki_content.page] ||= []
+        data[digest.wiki_content.project][digest.wiki_content.page].push digest.wiki_content
+        digest.destroy
+      end
     end
     
     WikiDigestMailer.wiki_digest(user, data).deliver unless data.empty?
